+++ src/cli/commands/public-cmds/checkout-cmd.ts
@@ -34,6 +34,16 @@ export default class Checkout extends Command {
     ['', 'existing', 'relevant for lanes. checkout only components in a lane that exist in the workspace'],
     ['v', 'verbose', 'showing verbose output for inspection'],
     ['', 'skip-npm-install', 'do not install packages of the imported components'],
+    [
+      '',
+      'ignore-package-json',
+      'do not generate package.json for the imported component(s). (it automatically enables skip-npm-install and save-dependencies-as-components flags)'
+    ],
+    [
+      '',
+      'conf [path]',
+      'write the configuration file (bit.json) and the envs configuration files (use --conf without path to write to the default dir)'
+    ],
     ['', 'ignore-dist', 'do not write dist files (when exist)']
   ];
   loader = true;
@@ -50,9 +60,11 @@ export default class Checkout extends Command {
       lane = false,
       verbose = false,
       skipNpmInstall = false,
-      ignoreDist = false,
       existing = false,
-      newLaneName
+      newLaneName,
+      ignorePackageJson = false,
+      conf,
+      ignoreDist = false
     }: {
       interactiveMerge?: boolean;
       ours?: boolean;
@@ -63,9 +75,11 @@ export default class Checkout extends Command {
       lane?: boolean;
       verbose?: boolean;
       skipNpmInstall?: boolean;
-      ignoreDist?: boolean;
       existing?: boolean;
       newLaneName?: string;
+      ignorePackageJson?: boolean;
+      conf?: string;
+      ignoreDist?: boolean;
     }
   ): Promise<ApplyVersionResults> {
     const checkoutProps: CheckoutProps = {
@@ -78,8 +92,13 @@ export default class Checkout extends Command {
       skipNpmInstall,
       existingOnWorkspaceOnly: existing,
       ignoreDist,
-      newLaneName
+      newLaneName,
+      ignorePackageJson,
+      writeConfig: !!conf
     };
+    if (typeof conf === 'string') {
+      checkoutProps.configDir = conf;
+    }
     return checkout(values, checkoutProps);
   }